Add a script 'test_filler.py' used when developing, implement a decent history rendered but still not perfect, need to improve it, bump to 0.4.0

+++ captains_log/cli/installer.py
@@ -1,29 +1,12 @@
 import os, shutil
 import click
 
+from captains_log.conf.install import install_app
 from captains_log.backend.init import init_database
 from captains_log.backend.models import CaptainsLogDatabase, Category, Entry
 
-def install_app(settings):
-    """
-    Common method to install app's stuff
-    """
-    if not os.path.exists(settings['config_dir']):
-        click.echo("* Creating the dedicated directory to: {0}".format(settings['config_dir']))
-        os.makedirs(settings['config_dir'])
-    else:
-        click.echo("* Fascinating, the dedicated directory allready exists at: {0}".format(settings['config_dir']))
-    
-    if not os.path.exists(settings['database_filepath']):
-        click.echo("* We need to create the database to: {0}".format(settings['database_filepath']))
-        init_database(settings, naive=True)
-        CaptainsLogDatabase.create_tables([Category, Entry])
-    else:
-        click.echo("* It is illogical, the database allready exists at: {0}".format(settings['database_filepath']))
-
-
 @click.command()
-@click.option('--force', default=False, help="Don't prompt user to ask for install")
+@click.option('--force/--no-force', default=False, help="Don't prompt user to ask for install")
 @click.pass_context
 def install_command(context, force=False):
     """
@@ -33,7 +16,7 @@ def install_command(context, force=False):
     
     if not force:
         click.confirm('Do you want to continue for installing my needed stuff?', default=True, abort=True)
-        click.echo("Make it so!")
+    click.echo("Make it so!")
     
     install_app(settings)
 
@@ -41,14 +24,16 @@ def install_command(context, force=False):
 
 
 @click.command()
+@click.option('--force/--no-force', default=False, help="Don't prompt user to ask for reset")
 @click.pass_context
-def reset_command(context):
+def reset_command(context, force):
     """
     To restart with a new install
     """
     settings = context.obj['settings']
     
-    click.confirm('Captain, this will reset your install and erase all your logs, are you sure ?', abort=True)
+    if not force:
+        click.confirm('Captain, this will reset your install and erase all your logs, are you sure ?', abort=True)
     click.echo("Setting Phasers on kill!")
     
     # Remove previous config
